Revert back to pending queues might have at least one record.
From PR comment:
Historically, queues have been "over-estimations", so for example a task
being in the pending queue meant that the task might actually be
pending. But a task that was not in the pending queue was definitely not
pending. This is a pretty common distributed-algorithm design, but was
necessary with azure queues since it's impossible to delete a message
from a queue - or know if a message exists in a queue - until that
message is visible and comes to the top of the queue.